+++ tests/test_optimizers.py
@@ -5,7 +5,6 @@ from torch import nn
 
 from pytorch_optimizer import GSAM, SAM, CosineScheduler, Lookahead, PCGrad, ProportionScheduler, load_optimizer
 from pytorch_optimizer.base.exception import NoClosureError, ZeroParameterSizeError
-from pytorch_optimizer.optimizer.shampoo_utils import BlockPartitioner, PreConditioner
 from tests.constants import ADAMD_SUPPORTED_OPTIMIZERS, ADAPTIVE_FLAGS, OPTIMIZERS, PULLBACK_MOMENTUM
 from tests.utils import (
     MultiHeadLogisticRegression,
@@ -27,8 +26,8 @@ def test_f32_optimizers(optimizer_fp32_config):
     optimizer_class, config, iterations = optimizer_fp32_config
 
     optimizer_name: str = optimizer_class.__name__
-    if optimizer_name == 'Nero' and 'constraints' not in config:
-        pytest.skip(f'skip {optimizer_name} w/o constraints')
+    if (optimizer_name == 'Nero' and 'constraints' not in config) or (optimizer_name == 'AliG'):
+        pytest.skip(f'skip {optimizer_name} w/ {config}')
 
     optimizer = optimizer_class(model.parameters(), **config)
 
@@ -83,6 +82,7 @@ def test_sam_optimizers(adaptive, optimizer_sam_config):
         or (optimizer_class.__name__ == 'DAdaptAdam')
         or (optimizer_class.__name__ == 'DAdaptSGD')
         or (optimizer_class.__name__ == 'Apollo' and 'weight_decay_type' in config)
+        or (optimizer_class.__name__ == 'AliG')
     ):
         pytest.skip(f'Skip {optimizer_class.__name__} w/ {config}')
 
@@ -305,21 +305,3 @@ def test_scalable_shampoo_pre_conditioner(pre_conditioner_type):
     loss_fn(model(x_data), y_data).backward()
 
     optimizer.step()
-
-
-def test_pre_conditioner():
-    var = torch.zeros((1024, 128))
-    grad = torch.zeros((1024, 128))
-
-    pre_conditioner = PreConditioner(var, 0.9, 0, 128, 8192, True, 1e-6, use_svd=False)
-    pre_conditioner.add_statistics(grad)
-    pre_conditioner.compute_pre_conditioners()
-
-
-def test_block_partitioner():
-    var = torch.zeros((2, 2))
-    target_var = torch.zeros((1, 1))
-
-    partitioner = BlockPartitioner(var, block_size=2, pre_conditioner_type=0)
-    with pytest.raises(ValueError):
-        partitioner.partition(target_var)
